move power transform logic to forecasting and make transform pub
