+++ crates/augurs-forecaster/src/lib.rs
@@ -3,12 +3,15 @@
 mod data;
 mod error;
 mod forecaster;
+mod power_transforms;
 pub mod transforms;
 
 pub use data::Data;
 pub use error::Error;
 pub use forecaster::Forecaster;
 pub use transforms::Transform;
-pub(crate) use transforms::Transforms;
+pub use transforms::Transforms;
+pub use power_transforms::optimize_lambda;
+
 
 type Result<T> = std::result::Result<T, Error>;
