+++ crates/augurs-forecaster/src/transforms.rs
@@ -7,15 +7,21 @@ use augurs_core::{
     Forecast,
 };
 
+/// Transforms and Transform implementations.
+/// 
+/// The `Transforms` struct is a collection of `Transform` instances that can be applied to a time series.
+/// The `Transform` enum represents a single transformation that can be applied to a time series.
 #[derive(Debug, Default)]
-pub(crate) struct Transforms(Vec<Transform>);
+pub struct Transforms(Vec<Transform>);
 
 impl Transforms {
-    pub(crate) fn new(transforms: Vec<Transform>) -> Self {
+    /// create a new `Transforms` instance with the given transforms.
+    pub fn new(transforms: Vec<Transform>) -> Self {
         Self(transforms)
     }
 
-    pub(crate) fn transform<'a, T>(&'a self, input: T) -> Box<dyn Iterator<Item = f64> + 'a>
+    /// Apply the transformations to the given time series.
+    pub fn transform<'a, T>(&'a self, input: T) -> Box<dyn Iterator<Item = f64> + 'a>
     where
         T: Iterator<Item = f64> + 'a,
     {
@@ -24,7 +30,8 @@ impl Transforms {
             .fold(Box::new(input) as _, |y, t| t.transform(y))
     }
 
-    pub(crate) fn inverse_transform(&self, forecast: Forecast) -> Forecast {
+    /// Apply the inverse transformations to the given forecast.
+    pub fn inverse_transform(&self, forecast: Forecast) -> Forecast {
         self.0
             .iter()
             .rev()
@@ -46,6 +53,8 @@ pub enum Transform {
     Logit,
     /// Log transform.
     Log,
+    /// Box-Cox transform.
+    BoxCox(f64),
 }
 
 impl Transform {
@@ -81,7 +90,16 @@ impl Transform {
         Self::Log
     }
 
-    pub(crate) fn transform<'a, T>(&'a self, input: T) -> Box<dyn Iterator<Item = f64> + 'a>
+
+    /// Create a new Box-Cox transform.
+    /// 
+    /// This transform applies the Box-Cox transformation to each item.
+    pub fn boxcox(lambda: f64) -> Self {
+        Self::BoxCox(lambda)
+    }
+
+    /// Apply the transformation to the given time series.
+    pub fn transform<'a, T>(&'a self, input: T) -> Box<dyn Iterator<Item = f64> + 'a>
     where
         T: Iterator<Item = f64> + 'a,
     {
@@ -90,10 +108,12 @@ impl Transform {
             Self::MinMaxScaler(params) => Box::new(input.min_max_scale(params.clone())),
             Self::Logit => Box::new(input.logit()),
             Self::Log => Box::new(input.log()),
+            Self::BoxCox(lambda) => Box::new(input.boxcox(*lambda)),
         }
     }
 
-    pub(crate) fn inverse_transform<'a, T>(&'a self, input: T) -> Box<dyn Iterator<Item = f64> + 'a>
+    /// Apply the inverse transformation to the given time series.
+    pub fn inverse_transform<'a, T>(&'a self, input: T) -> Box<dyn Iterator<Item = f64> + 'a>
     where
         T: Iterator<Item = f64> + 'a,
     {
@@ -102,10 +122,12 @@ impl Transform {
             Self::MinMaxScaler(params) => Box::new(input.inverse_min_max_scale(params.clone())),
             Self::Logit => Box::new(input.logistic()),
             Self::Log => Box::new(input.exp()),
+            Self::BoxCox(lambda) => Box::new(input.inverse_boxcox(*lambda)),
         }
     }
 
-    pub(crate) fn inverse_transform_forecast(&self, mut f: Forecast) -> Forecast {
+    /// Apply the inverse transformations to the given forecast.
+    pub fn inverse_transform_forecast(&self, mut f: Forecast) -> Forecast {
         f.point = self.inverse_transform(f.point.into_iter()).collect();
         if let Some(mut intervals) = f.intervals.take() {
             intervals.lower = self
@@ -371,6 +393,83 @@ trait ExpExt: Iterator<Item = f64> {
 
 impl<T> ExpExt for T where T: Iterator<Item = f64> {}
 
+
+/// Returns the Box-Cox transformation of the given value.
+/// Assumes x > 0.
+pub fn box_cox(x: f64, lambda: f64) -> f64 {
+    if lambda == 0.0 {
+        x.ln()
+    } else {
+        (x.powf(lambda) - 1.0) / lambda
+    }
+}
+
+/// An iterator adapter that applies the Box-Cox transformation to each item.
+#[derive(Clone, Debug)]
+struct BoxCox<T> {
+    inner: T,
+    lambda: f64,
+}
+
+impl<T> Iterator for BoxCox<T>
+where
+    T: Iterator<Item = f64>,
+{
+    type Item = f64;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.inner.next().map(|x| box_cox(x, self.lambda))
+    }
+}
+
+trait BoxCoxExt: Iterator<Item = f64> {
+    fn boxcox(self, lambda: f64) -> BoxCox<Self>
+    where
+        Self: Sized,
+    {
+        BoxCox { inner: self, lambda }
+    }
+}
+
+impl<T> BoxCoxExt for T where T: Iterator<Item = f64> {}
+
+/// Returns the inverse Box-Cox transformation of the given value.
+fn inverse_box_cox(y: f64, lambda: f64) -> f64 {
+    if lambda == 0.0 {
+        y.exp()
+    } else {
+        (y * lambda + 1.0).powf(1.0 / lambda)
+    }
+}
+
+/// An iterator adapter that applies the inverse Box-Cox transformation to each item.
+#[derive(Clone, Debug)]
+struct InverseBoxCox<T> {
+    inner: T,
+    lambda: f64,
+}
+
+impl<T> Iterator for InverseBoxCox<T>
+where
+    T: Iterator<Item = f64>,
+{
+    type Item = f64;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.inner.next().map(|y| inverse_box_cox(y, self.lambda))
+    }
+}
+
+trait InverseBoxCoxExt: Iterator<Item = f64> {
+    fn inverse_boxcox(self, lambda: f64) -> InverseBoxCox<Self>
+    where
+        Self: Sized,
+    {
+        InverseBoxCox { inner: self, lambda }
+    }
+}
+
+impl<T> InverseBoxCoxExt for T where T: Iterator<Item = f64> {}
+
+
 #[cfg(test)]
 mod test {
     use augurs_testing::{assert_all_close, assert_approx_eq};
@@ -501,5 +600,23 @@ mod test {
         assert_approx_eq!(params.data_max, 3.0);
         assert_approx_eq!(params.scaled_min, 0.0);
         assert_approx_eq!(params.scaled_max, 1.0);
+    }
+
+    #[test]
+    fn boxcox_test() {
+        let data = vec![1.0, 2.0, 3.0];
+        let lambda = 0.5;
+        let expected = vec![0.0, 0.8284271247461903, 1.4641016151377544];
+        let actual: Vec<_> = data.into_iter().boxcox(lambda).collect();
+        assert_all_close(&expected, &actual);
+    }
+
+    #[test]
+    fn inverse_boxcox_test() {
+        let data = vec![0.0, 0.5_f64.ln(), 1.0_f64.ln()];
+        let lambda = 0.5;
+        let expected = vec![1.0, 0.426966072919605, 1.0];
+        let actual: Vec<_> = data.into_iter().inverse_boxcox(lambda).collect();
+        assert_all_close(&expected, &actual);
     }
 }
