--- crates/augurs-transformations/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "augurs-transformations"
-version.workspace = true
-authors.workspace = true
-documentation.workspace = true
-repository.workspace = true
-license.workspace = true
-edition.workspace = true
-keywords = [
-  "transformations",
-  "time-series",
-]
-description = "Time series data transformations"
-
-[lib]
-bench = false
-
-[dependencies]
-augurs-core.workspace = true
-itertools.workspace = true
-thiserror.workspace = true
-argmin = "0.10.0"
-
-[dev-dependencies]
-augurs.workspace = true
-augurs-testing.workspace = true
-
-[lints]
-workspace = true
