--- crates/augurs-transformations/README.md
@@ -1,11 +1,0 @@
-# Time series data transformations
-
-Transformation pipeline for time series data
-
-## Example
-
-```rust
-```
-
-## Credits
-
